feat: create template for full status parser
